feat(U-6947614384): Handle Service Definition Changes
The update introduces a ServiceDefinitionConverter class to manage the conversion between the ServiceDefinition model and its JSON representation. This simplifies the handling of ServiceDefinition instances by replacing manual JSON (de)serialization with automatic JPA conversions. With this update, the ServiceDefinition field in ServiceDefinitionEntity is directly manipulated as a ServiceDefinition instance, eliminating the need for explicit conversions.

Signed-off-by: Roman Naglič <naglicr@tolam.io>