feat(U-6962838906): Return Jurisdiction ID and Jurisdiction Name To Front End
Implemented a new RemoteHost entity to better represent the relationship between jurisdictions and hosts. This change modified the Jurisdiction model to include a set of RemoteHosts instead of a single hostName. The JurisdictionService class was updated to find jurisdictions using these new RemoteHost entities. The corresponding tests and repository methods were also revised for these changes.

Signed-off-by: Roman Naglič <naglicr@tolam.io>